src/requirements.txt
aiohttp==3.6.2
aiomysql==0.0.20
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.8
cycler==0.10.0
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fastapi==0.52.0
Flask==1.1.1
h11==0.9.0
idna==2.9
idna-ssl==1.1.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
multidict==4.7.5
mysql-connector==2.2.9
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.2
patsy==0.5.1
pickleshare==0.7.5
Pillow==7.0.0
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==3.0.3
psutil==5.7.0
pycparser==2.19
pydantic==1.4
pyecharts==1.6.2
Pygments==2.5.2
PyMySQL==0.9.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
requests==2.23.0
scikit-learn==0.22.2
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
simplejson==3.17.0
six==1.14.0
sklearn==0.0
SQLAlchemy==1.3.13
starlette==0.13.2
statsmodels==0.11.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
typing-extensions==3.7.4.1
urllib3==1.25.8
uvicorn==0.11.3
wcwidth==0.1.8
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
yarl==1.4.2
zipp==3.0.0
